{topost} scsi: hisi_sas: relocate the code of clearing itct and freeing device
In certain scenarios we may just want to clear the itct for a device,
and not free other resources like the SATA bitmap using in v2 hw.

To facilitate this, this patch relocate the code of clearing itct
from free_device() to an new hw interface clear_itct(). Then for
some hw, we should not realize free_device() if there's nothing
left to do for it.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>