{noup} crypto: wrapdrive: wrapdrive framework
This is the wrapdrive framwork implmentation
It provides a set of helper function to the accelerator driver
to export the device interface to the user space via vfio-mdev framework.

Therefore the userspace can get the capabilities of the hardware and
required a queue, which is wrapped as a vfio-mdev device. So the user_driver
can access the device via standard vfio interface.

Signed-off-by: Zaibo Xu <xuzaibo@huawei.com>
Signed-off-by: Kenneth Lee <liguozhu@hisilicon.com>
Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>