{noup} wrapdrive: The wrapdrive user interface
Interfaces in the user space:
	int wd_request_queue(struct wd_queue *q, struct wd_capa *capa);
        void wd_release_queue(struct wd_queue *q);

        int wd_send(struct wd_queue *q, void *req);
        int wd_recv(struct wd_queue *q, void **req);
        int wd_send_sync(struct wd_queue *q, void *req);
        int wd_recv_sync(struct wd_queue *q, void **req);

The wd_queue may not directly refer to a hardware queue. It refers only to a
handle, that the user application can get help from it by sending request. So
the "user driver" may just reuse the same hardware queue for more than one
wd_queue. It is all depend on the driver developers.

The hardware accelerator will need to access the user memory space in many
cases. To share the memory with it, the user application can request queues
with capabilities with WD_CAPA_SHARE_ALL or it can share the memory with
explicit calls:

        int wd_mem_share(struct wd_queue *q, const void *addr, size_t size, int flags);
        void wd_mem_unshare(struct wd_queue *q, const void *addr, size_t size);

The WD_CAPA_SHARE_ALL is not imperative to all hardware types.

Signed-off-by: Zaibo Xu <xuzaibo@huawei.com>
Signed-off-by: Kenneth Lee <liguozhu@hisilicon.com>