{topost} scsi: hisi_sas: do link reset for some interrupts in CHL_INT2
We should do link reset of PHY when identify timeout or STP link
timeout. They are internal events of SOC and are notified to driver
through interrupts of CHL_INT2.

Besides, we should add an delay work to do link reset as it needs sleep.
So, this patch add an new PHY event HISI_PHYE_LINK_RESET for this.

Notes: v2 HW doesn't report the event of STP link timeout. So, we
only need to handle event of identify timeout for v2 HW.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>