{topost} scsi: hisi_sas: fix SAS_QUEUE_FULL problem while running io
This patch fix SAS_QUEUE_FULL problem. The test situation is
close port while running io.

In sas_eh_handle_sas_errors(), SCSI EH will free sas_task of the devcie
if lldd_I_T_nexus_reset() return TMF_RESP_FUNC_COMPLETE or -ENODEV.
But in our SAS driver, we only free slots of the device when the return
value is TMF_RESP_FUNC_COMPLETE. So if the return value is -ENODEV,
the slot resource will not free any more.

As an solution, we should also free slots of the device in
lldd_I_T_nexus_reset() if the return value is -ENODEV.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>