{topost} perf utils: add support for pmu events vendor sub-directory
For some architectures (like arm), it is required to support a
vendor sub-directory and not put all the jsons for a specific
vendor in the same folder.

This is because all the events for the same vendor will be placed
in the same pmu events table, which may cause conflict. This conflict
would be in the instance that a vendor's custom implemented events do
have the same meaning on different platforms, so events in the pmu
table would conflict. In addition, per list command may show events
which are not even supported for a given platform.

This patch adds support for a arch/vendor/platform directory
hierarchy, while maintaining backwards-compatibility for classic
arch/platform structure. In this, each platform would always have
its own pmu events table.

In generated file pmu_events.c, each platform table name is in
the format pme{_vendor}_platform, like this:

struct pmu_events_map pmu_events_map[] = {
{
	.cpuid = "0x00000000420f5160",
	.version = "v1",
	.type = "core",
	.table = pme_cavium_thunderx2
},
{
	.cpuid = 0,
	.version = 0,
	.type = 0,
	.table = 0,
},
};

In addition to adding vendor sub-directory support, the Cavium
ThunderX2 json is relocated to a new appropriate location.

Signed-off-by: John Garry <john.garry@huawei.com>