{topost} perf utils: add support for arch recommended events
For some architectures (like arm), there are architecure-defined
recommended defined. Vendors are not obliged to follow the
recommendation and may implement their own pmu events for a
specific event code.

This patch adds support for parsing recommended events from
arch-defined recommended jsons, and fixing up vendor events
when they have implemented these events.

A new step is added to the pmu events parsing to fix up the
vendor events with the arch-recommended events.

In the vendor json, to specify that the recommended event is
supported, only the event code is added to the json entry - no
other event elements should be added. If the "BriefDescription"
is missing, then it is implied that the recommended event
has been implemented, like this:
[
    {
        "EventCode": "0x40",
    },

]

If the "BriefDescription" is present, then it is implied that the
vendor has implemented their own custom event, and there is no
fixup.

Signed-off-by: John Garry <john.garry@huawei.com>