{topost} scsi: hisi_sas: add internal abort dev in some places.
We should do internal abort dev before TMF_ABORT_TASK_SET and
TMF_LU_RESET. Because we may only have done internal abort for
single IO in the earlier part of SCSI EH process. Even the
internal abort to the single IO, we also don't know whether
it is successful.

Besides, we should release slots of the device in
hisi_sas_abort_task_set() if the abort is successful.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>