{fromlist v5} arm64: toplogy: add cluster_id and cluster_sibling in struct cpu_topology
We also need to parse the cluster infomation from PPTT table, becuase we
need that information to build the MC sched_domain. So we add
'cluster_id' and 'cluster_sibling' in struct cpu_topology.
'core_sibling' is redefined to represent cores in a cluster.
'cluster_sibling' is used to represent cores in a package. Since the ARM64
sched_domain building framework uses struct cpu_topology.core_sibling
to build the MC sched_domain, so we can let cores of a cluster in one
sched_domain, rather than cores of a package.

Signed-off-by: Xiongfeng Wang <wangxiongfeng2@huawei.com>
Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>