{topost} scsi: libsas: always unregister the old device if going to discover new
If we went into sas_rediscover_dev() the attached_sas_addr was already
insured not to be 0. So it's unnecessary to check if the
attached_sas_addr is 0.

And although if the sas address is not changed, we always have to
unregister the old device when we are going to register a new one. We
cannot just leave the device there and bring up the new.

Signed-off-by: Jason Yan <yanaijie@huawei.com>