{topost} scsi: libsas: do not consider sata device broadcast as flutter
The ata device do not have a real sas address. If a ata device is
replaced with another one, the sas address is the same. Now libsas treat
this senario as flutter and do not delete the old one and discover the
new one.

Do not treat ata device broadcast as flutter. Always delete the old
device and probe a new one.

Signed-off-by: Jason Yan <yanaijie@huawei.com>