iommu: Add a mm_exit callback for device drivers
When an mm exits, we need to make sure that devices attached to it stop
issuing transactions with its PASID. Let device drivers register a
callback to be notified on mm exit.

At the moment the callback is set on the domain like the fault handler,
because we don't have a structure available for IOMMU masters. This can
become problematic if different devices in a domain are managed by
distinct device drivers (for example multiple devices in the same group).
The problem is the same for the fault handler, so we'll probably fix them
all at once.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>