iommu/sva: Bind and unbind process address spaces to and from devices
Shared Virtual Addressing (SVA) provides a way for device drivers to bind
process address spaces to devices. This requires the IOMMU to support page
table sharing, and requires the system to support I/O Page Faults (IOPF)
and Process Address Space ID (PASID). When all of these are available, DMA
can directly target virtual addresses of a process. A PASID is allocated
for each process, that is configured in the device in an implementation-
specific way.

Add bind() and unbind() operations to the IOMMU API. Device drivers can
use them to share process page tables with their device. bind_group() is
provided for VFIO's convenience, as it needs to provide a coherent
interface on containers. Device drivers will most likely want to use
bind_device(), which binds a single device in the group.

Regardless of the IOMMU group or domain a device is in, device drivers
should call bind() for each device that will use the PASID.

This patch only add skeletons for the device driver API.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>