ACPI / PPTT: Add helper to validate cache nodes from an offset [dead]
The MPAM table holds a PPTT-offset that describes a cache. Add a
helper acpi_pptt_validate_cache_node(), which checks this offset
really is in the PPTT, on a node boundary, and that node really
is a cache.

With this the MPAM probe code can go poking through struct cacheinfo
looking for a level with a matching firmware_node.

Signed-off-by: James Morse <james.morse@arm.com>