{fromlist v1} scsi: hisi_sas: increase timer expire of internal abort task
The 110ms is not enough for the timer expire of internal abort task.
Because internal abort task could be blocked in HW, if the HW is
retrying to set up link. The internal abort task will be executed
only when the retry process finished.

The maximum time is 5s for the retry of setting up link. So, the timer
expire should be more than 5s. This patch increases it from 110ms to 6s.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>