{fromtree} ACPI / scan: rename acpi_is_serial_bus_slave() to widen use
Currently the ACPI scan has special handling for serial bus
slaves, in that it makes it the responsibility of the the slave
device's parent to enumerate the device.

To support in future other types of slave devices which require
the same special handling, but where the bus is not strictly a
serial bus - such as devices on the HiSilicon LPC controller bus -
rename acpi_is_serial_bus_slave() to
acpi_device_enumeration_by_parent(), so that the name can fit the
wider purpose.

Associated device flag acpi_device_flags.serial_bus_slave is also
renamed to .enumeration_by_parent.

Signed-off-by: John Garry <john.garry@huawei.com>
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>