{topost} scsi: hisi_sas: check IPTT is valid before using it for v3 hw
There is an SOC bug of v3 hw development version. When AXI error happen,
hw may return an abnormal CQ that IPTT value is 0xffff. This will cause
out-of-bounds access of slot_info.

This patch add an check of IPTT in cq_tasklet_v3_hw(), and discard
invalid slot. This workaround scheme is just to enhance fault-tolerant
ability of our driver. So, we will apply this scheme for all version
of v3 hw, although release version has fixed this SOC bug.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>