{topost} scsi: hisi_sas: block IO when host controller is during reset
When host controller is resetting, hisi_sas_task_exec() just returns
-EINVAL directly in current driver. This may cause SMP IO failure, and
the failure SMP IO won't be going to retry. This is unacceptable for
libsas.

This patch uses hisi_hba->sem to block IO during host reset. If IO is
running in softirq, We can't grap semaphore. For such IOs, we just
retain the original handling way . This is ok. Because libsas never
run SMP IO in softirq.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>