{topost} scsi: hisi_sas: remove redundant handling to event95 for v3 hw
Event95 is used as DFX purpose，and its interrupt mask is bit31 of
ENT_INT_SRC_MSK3. We have disabled it for v3 hw, and don't need to
handle it in int_chnl_int_v3_hw() anymore.

This patch is to remove this redundant handling to event95 for v3 hw.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>