{topost} net: hns3: fix for registering client problem
If the client is not registered in hnae3_register_client, it will
cause an error in the function hnae3_unregister_client to call the
list entry checking function __list_del_entry_valid.

This patch fixes this problem by initializing the node member which
contained in struct hnae3_client as an empty list.

Signed-off-by: Xi Wang <wangxi11@huawei.com>
Signed-off-by: Peng Li <lipeng321@huawei.com>