{topost} scsi: hisi_sas: adjust IO reject period of host reset
After soft_reset(), it shouldn't be allowed to send IOs before
PHYs up and port id refresh. Because if they are not done, IOs
can't be successful. This patch is to fix this issue.

Besides, put up(&hisi_hba->sem) ahead to recover the IOs quickly,
that have been blocked by this semaphore.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>