{topost} scsi: hisi_sas: release all remain resources in clear nexus ha
In host reset, we use TMF or soft-reset to re-init device, and if
success, we will release all LLDD resources of this device. If the
init fails - maybe because the device was removed or link has
not come up - then do not release the LLDD resources, but rather
rely on SCSI EH to handle the timeout for these resources later on.

But if clear nexus ha calls host reset, which is the last effort of
SCSI EH, we should release all LLDD remain resources. Because SCSI
EH will release all tasks after clear nexus ha.

Before release, we do I_T nexus reset to try to clear target remain IOs.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>