{topost} scsi: hisi_sas: read native max address in dev found
This patch adds an command of reading native max address in dev found.
The reason is that SAMSUNG SATA SSDs need 3-5s to response the command
if the disk is just powered on. This command is used in process of
ATA device probe and will hold time of SCSI EH.

To workaroud this issue, we add the command, and make the time waiting
in dev found function. Besides, if the command is failed, we do internal
abort and soft-reset to clean it.

Signed-off-by: Xiaofei Tan <tanxiaofei@huawei.com>
Signed-off-by: John Garry <john.garry@huawei.com>