{topost} scsi: hisi_sas: relocate a check to avoid null pointer
If SMP IO is timeout when device gone, it will call hisi_sas_abort_task to
recover. We have a check in function hisi_sas_abort_task to avoid the
situation. But before it we refer to domain_device, and actually at that time
domain_device may be freed, so it is a risk to refer to null pointer in line
"struct hisi_hba *hisi_hba = dev_to_hisi_hba(task->dev)".
So relocate the check and refer to dev_to_hisi_hba after the check.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>