{topost} scsi: hisi_sas: config ATA de-reset as a constraint command
As a unconstraint command, the command can be sent to SATA disk even if
SATA disk status is in BUSY, ERR or DRQ status.
If ata reset is successful but ATA de-reset is failed, then it will try
de-reset currently. If de-reset retried is successful, we think it is okay
to probe device but actually it is still in Err status. According to Soc guy,
we need to retry ATA reset and de-reset if de-reset is failed.
So we config ATA de-reset as a contrain command, if ATA de-reset is failed,
then ATA de-reset retried will also be failed. Then it will retry ATA reset
and ATA de-reset again.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>