{fromlist v3} ACPI / APEI: Move the estatus queue code up, and under its own ifdef
To support asynchronous NMI-like notifications on arm64 we need to use
the estatus-queue. These patches refactor it to allow multiple APEI
notification types to use it.

First we move the estatus-queue code higher in the file so that any
notify_foo() handler can make use of it.

This patch moves code around ... and makes the following trivial change:
Freshen the dated comment above ghes_estatus_llist. printk() is no
longer the issue, its the helpers like memory_failure_queue() that
still aren't nmi safe.

Signed-off-by: James Morse <james.morse@arm.com>
Reviewed-by: Punit Agrawal <punit.agrawal@arm.com>

Notes for cover letter:
ghes.c has three things all called 'estatus'. One is a pool of memory
that has a static size, and is grown/shrunk when new NMI users are
allocated.
The second is the cache, this holds recent notifications so we can
suppress notifications we've already handled.
The last is the queue, which hold data from NMI notifications (in pool
memory) that can't be handled immediatly.
Reviewed-by: Borislav Petkov <bp@suse.de>