{fromlist v1} HISI LPC: Reference static MFD cells for ACPI support
Currently for ACPI support the driver models the host as
an MFD. For a device connected to the LPC bus, we dynamically
create an MFD cell for that device, configuring the cell
name and ACPI match parameters manually. This makes supporting
named devices and also special setup handling for certain devices
awkward, as we would need to introduce some special ACPI device
handling according to device HID.

To avoid this, create reference static MFD cells for known
devices, so when adding an MFD cell we can fix the cell
configuration. Also, in future, we can add a setup callback
function for the cell.

For now, only the IPMI code cell is added.

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Wei Xu <xuwei5@hisilicon.com>