{fromlist v1} HISI LPC: Add ACPI UART support
On the Huawei D03 development board the system UART is
the UART connected on the LPC bus.

The profile for the device driver required for this HW
is as follows:
- platform driver
- 16550
- ACPI support
- polling mode support
- IO space support

In principle we should use the PNP driver (8250_dw.c) for
8250-devices with ACPI FW. However since this driver does
not support PNP devices, and modifying the PNP core code
to support it is not worth the effort, use the generic
8250 isa driver.

For this, we need to setup the MFD cell platform data for a
serial 8250 Port.

In addition to this change, we also make the following
changes:
- set ACPI device as enumerated. This fixes a conflict in
  serdev config for the UART
- use device ID auto for creating the child platform device
  to avoid conflict with platform devices created for
  serial8250_isa_devs from 8250_core.c

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Wei Xu <xuwei5@hisilicon.com>